Abstract darksky service to its own file
