+++ app/routes/api/v1/forecasts.js
@@ -1,5 +1,6 @@
 const setup = require('./index')
 const formatter = require('../../../formatters/forecastFormatter')
+const darksky = require('../../../services/darkskyService')
 const router = setup.router
 const fetch = setup.fetch
 const findKey = setup.findByKey
@@ -9,12 +10,12 @@ function googleGeocode (address) {
     .catch((error) => console.error({ error }))
 }
 
-function darkskyForecast (coordinates) {
-  const lat = coordinates.lat
-  const long = coordinates.lng
-  return fetch(`https://api.darksky.net/forecast/${process.env.DARKSKY_API_KEY}/${lat},${long}`)
-    .catch((error) => console.error({ error }))
-}
+// function darkskyForecast (coordinates) {
+//   const lat = coordinates.lat
+//   const long = coordinates.lng
+//   return fetch(`https://api.darksky.net/forecast/${process.env.DARKSKY_API_KEY}/${lat},${long}`)
+//     .catch((error) => console.error({ error }))
+// }
 
 router.get('/', (request, response) => {
   (async () => {
@@ -24,13 +25,13 @@ router.get('/', (request, response) => {
 
     if (userId) {
       var coordinates = (await googleGeocode(address).then(response => response.json())).results[0].geometry.location
-      var darksky = (await darkskyForecast(coordinates).then(response => response.json()))
+      var darkdata = (await darksky(coordinates).then(response => response.json()))
       var forecast = {}
 
       forecast.location = address
-      forecast.currently = await formatter.formatCurrently(darksky)
-      forecast.hourly = await formatter.formatHourly(darksky)
-      forecast.daily = await formatter.formatDaily(darksky)
+      forecast.currently = await formatter.formatCurrently(darkdata)
+      forecast.hourly = await formatter.formatHourly(darkdata)
+      forecast.daily = await formatter.formatDaily(darkdata)
 
       response.status(200).send(forecast)
     } else {
